NHL player population improved
Setting up the NHL API to run through each NHL team and grab its active roster. Success so far, but currently it replaces the list repeatedly, leaving only the last team done in the data set, which appears to be SJS. Also just printing to a json currently, should be adjusted to save to a proper database.